refactor(scanner): extract ore scanner state and OreColorMapping from Eln.java into OreScannerManager
Move oreScannerConfig field, regenOreScannerFactors() method, and
OreColorMapping object out of Eln.java and PortableOreScannerItem.kt
into new mods.eln.ore.OreScannerManager. Pure refactoring, no behavioral changes.
