+++ src/main/java/mods/eln/client/ConnectionListener.java
@@ -8,6 +8,7 @@ import cpw.mods.fml.common.network.FMLNetworkEvent.ClientConnectedToServerEvent;
 import cpw.mods.fml.common.network.FMLNetworkEvent.ClientDisconnectionFromServerEvent;
 import mods.eln.Eln;
 import mods.eln.misc.Utils;
+import mods.eln.ore.OreScannerManager;
 import mods.eln.misc.UtilsClient;
 import net.minecraft.client.Minecraft;
 
@@ -27,7 +28,7 @@ public class ConnectionListener {
     @SubscribeEvent
     public void onConnectedToServerEvent(ClientConnectedToServerEvent event) {
         Utils.println("Connected to server " + FMLCommonHandler.instance().getEffectiveSide());
-        Eln.instance.regenOreScannerFactors();
+        OreScannerManager.regenOreScannerFactors();
 
         timer = 20;
         newConnection = true;
